Updated Readme to include Java implementation
I've created a java implementation of the IgushArray at https://github.com/StoneT2000/IgushArray-Java, which is still a work in progress, but the test code already shows the massive speed improvements in terms of insertion and removal.

The things missing are proper test code that go through all the logic, and good stat graphs! As well as some overridden methods that need to be implemented in order to be a one for one replacement for the ArrayList in Java.