Update README.md to include JS implementation
I wrote a JS implementation of the IgushArray and added it to the readme. It's also published on NPM. It's basic functionality is intact and is usable for majority of use cases (major functions such as push, pop, unshift, shift, sort all implemented), with some other methods built in arrays have to be implemented soon.

Again, it shows dramatic improvements in insertion and removal times, but some testing is somewhat difficult considering the way the V8 engine handles dynamic array resizes. For the most part, the IgushArray is just faster, with access times and slice times slower only on the magnitude of 10^-8 seconds.